+++ rest_framework_gis/fields.py
@@ -1,8 +1,9 @@
+import six  # TODO Remove this along with GeoJsonDict when support for python 2.6/2.7 is dropped.
 import json
 from collections import OrderedDict
 
 from django.contrib.gis.geos import GEOSGeometry, GEOSException
-from django.contrib.gis.gdal import OGRException
+from django.contrib.gis.gdal import GDALException
 from django.core.exceptions import ValidationError
 from django.utils.translation import ugettext_lazy as _
 from rest_framework.fields import Field, SerializerMethodField
@@ -25,11 +26,7 @@ class GeometryField(Field):
         if isinstance(value, dict) or value is None:
             return value
         # we expect value to be a GEOSGeometry instance
-        value.transform("EPSG: 4326")
-        return GeoJsonDict((
-            ('type', value.geom_type),
-            ('coordinates', value.coords),
-        ))
+        return GeoJsonDict(value.geojson)
 
     def to_internal_value(self, value):
         if value == '' or value is None:
@@ -41,8 +38,10 @@ class GeometryField(Field):
             value = json.dumps(value)
         try:
             return GEOSGeometry(value)
-        except (ValueError, GEOSException, OGRException, TypeError):
+        except (GEOSException):
             raise ValidationError(_('Invalid format: string or unicode input unrecognized as GeoJSON, WKT EWKT or HEXEWKB.'))
+        except (ValueError, TypeError, GDALException) as e:
+            raise ValidationError(_('Unable to convert to python object: {}'.format(str(e))))
 
     def validate_empty_values(self, data):
         if data == '':
@@ -55,10 +54,7 @@ class GeometrySerializerMethodField(SerializerMethodField):
         value = super(GeometrySerializerMethodField, self).to_representation(value)
         if value is not None:
             # we expect value to be a GEOSGeometry instance
-            return GeoJsonDict((
-                ('type', value.geom_type),
-                ('coordinates', value.coords),
-            ))
+            return GeoJsonDict(value.geojson)
         else:
             return None
 
@@ -68,6 +64,18 @@ class GeoJsonDict(OrderedDict):
     Used for serializing GIS values to GeoJSON values
     TODO: remove this when support for python 2.6/2.7 will be dropped
     """
+    def __init__(self, *args, **kwargs):
+        """
+        If a string is passed attempt to pass it through json.loads,
+        because it should be a geojson formatted string.
+        """
+        if args and isinstance(args[0], six.string_types):
+            try:
+                geojson = json.loads(args[0])
+                args = (geojson,)
+            except ValueError:
+                pass
+        super(GeoJsonDict, self).__init__(*args, **kwargs)
 
     def __str__(self):
         """